lastDigit

+++ variables_if_else_while/1-last_digit.c
@@ -12,7 +12,7 @@ int main(void)
 
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
-	int lastDigit = (n % 10);
+	long int lastDigit = (n % 10);
 
 	printf("Last digit of %d is", n);
 	if (lastDigit > 5)
